+++ csvtk/cmd/collapse.go
@@ -69,7 +69,7 @@ var collapseCmd = &cobra.Command{
 
 		fieldStr = fmt.Sprintf("%s,%s", fieldStr, vfieldStr)
 
-		fields, colnames, negativeFields, needParseHeaderRow := parseFields(cmd, fieldStr, config.NoHeaderRow)
+		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, config.NoHeaderRow)
 		var fieldsMap map[int]struct{}
 		var fieldsOrder map[int]int      // for set the order of fields
 		var colnamesOrder map[string]int // for set the order of fields