+++ csvtk/cmd/filter.go
@@ -73,7 +73,7 @@ var filterCmd = &cobra.Command{
 		threshold, err := strconv.ParseFloat(items[0][3], 64)
 		checkError(err)
 
-		fields, colnames, negativeFields, needParseHeaderRow := parseFields(cmd, fieldStr, config.NoHeaderRow)
+		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, config.NoHeaderRow)
 		var fieldsMap map[int]struct{}
 		if len(fields) > 0 {
 			fields2 := make([]int, len(fields))