+++ csvtk/cmd/filter2.go
@@ -95,7 +95,7 @@ Supported operators and types:
 
 		usingColname := true
 
-		fields, colnames, negativeFields, needParseHeaderRow := parseFields(cmd, fieldStr, config.NoHeaderRow)
+		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, config.NoHeaderRow)
 		var fieldsMap map[int]struct{}
 		if len(fields) > 0 {
 			usingColname = false