+++ csvtk/cmd/rename.go
@@ -55,7 +55,7 @@ var renameCmd = &cobra.Command{
 		}
 		names := getFlagCommaSeparatedStrings(cmd, "names")
 
-		fields, colnames, negativeFields, needParseHeaderRow := parseFields(cmd, fieldStr, config.NoHeaderRow)
+		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, config.NoHeaderRow)
 		var fieldsMap map[int]struct{}
 		if len(fields) > 0 {
 			fields2 := make([]int, len(fields))