+++ csvtk/cmd/sep.go
@@ -90,7 +90,7 @@ var sepCmd = &cobra.Command{
 		if fieldStr == "" {
 			checkError(fmt.Errorf("flag -f (--fields) needed"))
 		}
-		fields, colnames, negativeFields, needParseHeaderRow := parseFields(cmd, fieldStr, config.NoHeaderRow)
+		fields, colnames, negativeFields, needParseHeaderRow, _ := parseFields(cmd, fieldStr, config.NoHeaderRow)
 		if !(len(fields) == 1 || len(colnames) == 1) {
 			checkError(fmt.Errorf("only single field allowed"))
 		}