+++ CHANGES.md
@@ -1,3 +1,6 @@
+- [csvtk v0.21.1](https://github.com/shenwei356/csvtk/releases/tag/v0.21.1)
+[![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.21.1/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.21.1)
+    - `csvtk cut`: support selecting format `2-` for choose 2nd column to the end.
 - [csvtk v0.21.0](https://github.com/shenwei356/csvtk/releases/tag/v0.21.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.21.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.21.0)
     - new command: `csvtk nrow/ncol` for printing number of rows or columns.